+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Make dropdown element initially visible if `open` attribute is set
 - Export tabs element from `package.json` file
 
 ## [0.1.0] - 2024-01-06